Correções para não remover clientes e produtos relacionados com pedidos.

+++ SeusPedidos/Api/views/cliente.py
@@ -3,6 +3,7 @@ import json
 from django.http import HttpResponse
 from django.core import serializers
 from SeusPedidos.App.models.cliente import Cliente as ClienteModel
+from SeusPedidos.App.models.pedido import Pedido as PedidoModel
 from SeusPedidos.App.core.apiview import ApiView
 from SeusPedidos.App.form.cliente import ClienteForm
 
@@ -59,11 +60,15 @@ class Cliente(ApiView):
 
     def delete(self, request):
         id = request.GET.get('id')
-        try:
-            cliente = ClienteModel.objects.get(id=id)
-            cliente.delete()
-            result = self._apiresult.success(None)
-        except:
+        total = PedidoModel.objects.filter(cliente=id).count()
+        if (total == 0):
+            try:
+                cliente = ClienteModel.objects.get(id=id)
+                cliente.delete()
+                result = self._apiresult.success(None)
+            except:
+                result = self._apiresult.error(None)
+        else:
             result = self._apiresult.error(None)
         return HttpResponse(
             json.dumps(result)
